% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{aggr.ffms_predict}
\alias{aggr.ffms_predict}
\title{Access Aggregated Predictions}
\usage{
\method{aggr}{ffms_predict}(object, ...)
}
\arguments{
\item{object}{Object of class "ffms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
List containing aggregated mean and quantiles.
}
\description{
Extracts the aggregated predictions (mean and quantiles) from an FFMS prediction object.
}
\examples{
\donttest{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
aggr(pred)
}
}
