% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coef.fms_base}
\alias{coef.fms_base}
\title{Coefficients for FMS Model}
\usage{
\method{coef}{fms_base}(object, ...)
}
\arguments{
\item{object}{Object of class "fms_base".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of coefficients from the best model found.
}
\description{
Extracts coefficients from the best FMS model.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "fms_base")
coef(model)
}
