% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coef.mjmcmc}
\alias{coef.mjmcmc}
\title{Coefficients for MJMCMC Model}
\usage{
\method{coef}{mjmcmc}(object, ...)
}
\arguments{
\item{object}{Object of class "mjmcmc".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of coefficients from the best model found.
}
\description{
Extracts coefficients from the best MJMCMC model.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc")
coef(model)
}
