% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coef.mjmcmc_parallel}
\alias{coef.mjmcmc_parallel}
\title{Coefficients for MJMCMC Parallel Model}
\usage{
\method{coef}{mjmcmc_parallel}(object, ...)
}
\arguments{
\item{object}{Object of class "mjmcmc_parallel".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of coefficients from the best model found.
}
\description{
Extracts coefficients from the best MJMCMC parallel model.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc.parallel", cores = 1, runs = 2)
coef(model)
}
