% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.ffms_predict}
\alias{fitted.ffms_predict}
\title{Access Fitted Values}
\usage{
\method{fitted}{ffms_predict}(object, ...)
}
\arguments{
\item{object}{Object of class "ffms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of mean predictions.
}
\description{
Extracts the mean predictions from an FFMS prediction object (alias for mean).
}
\examples{
\donttest{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
fitted(pred)
}
}
