% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predquantiles.ffms_predict}
\alias{predquantiles.ffms_predict}
\title{Access Quantile Predictions}
\usage{
\method{predquantiles}{ffms_predict}(object, ...)
}
\arguments{
\item{object}{Object of class "ffms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Matrix of quantile predictions, or NULL if not available.
}
\description{
Extracts the quantile predictions from an FFMS prediction object.
}
\examples{
\donttest{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc")
pred <- predict(model, exoplanet[51:60, -1])
predquantiles(pred)
}
}
