% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.ffms_predict}
\alias{print.ffms_predict}
\title{Print FFMS Prediction Object}
\usage{
\method{print}{ffms_predict}(x, ...)
}
\arguments{
\item{x}{Object of class "ffms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a summary and returns NULL.
}
\description{
Displays a summary of an FFMS prediction object, including mean predictions and quantiles.
}
\examples{
\donttest{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
print(pred)
}
}
