% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fms_base}
\alias{print.fms_base}
\title{Print FMS Model Object}
\usage{
\method{print}{fms_base}(x, ...)
}
\arguments{
\item{x}{Object of class "fms_base".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of an FMS model object.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "fms_base")
print(model)
}
