% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fms_parallel}
\alias{print.fms_parallel}
\title{Print FMS Parallel Model Object}
\usage{
\method{print}{fms_parallel}(x, ...)
}
\arguments{
\item{x}{Object of class "fms_parallel".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of an FMS parallel model object.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "fms.parallel", cores = 1, runs = 2)
print(model)
}
