% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gmjmcmc_merged}
\alias{print.gmjmcmc_merged}
\title{Print GMJMCMC Merged Model Object}
\usage{
\method{print}{gmjmcmc_merged}(x, ...)
}
\arguments{
\item{x}{Object of class "gmjmcmc_merged".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of a GMJMCMC merged model object.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, 
method = "gmjmcmc.parallel", cores = 1, runs = 2, 
transforms = c("sigmoid"))

print(model)
}
