% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mjmcmc}
\alias{print.mjmcmc}
\title{Print MJMCMC Model Object}
\usage{
\method{print}{mjmcmc}(x, ...)
}
\arguments{
\item{x}{Object of class "mjmcmc".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of an MJMCMC model object.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc")
print(model)
}
