% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mjmcmc_parallel}
\alias{print.mjmcmc_parallel}
\title{Print MJMCMC Parallel Model Object}
\usage{
\method{print}{mjmcmc_parallel}(x, ...)
}
\arguments{
\item{x}{Object of class "mjmcmc_parallel".}

\item{...}{Additional arguments passed to summary method.}
}
\value{
Prints a summary of the model and returns NULL
}
\description{
Displays a concise summary of an MJMCMC parallel model object.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc.parallel", cores = 1, runs = 2)
print(model)
}
