% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.ffms_base}
\alias{residuals.ffms_base}
\title{Residuals for FFMS Model}
\usage{
\method{residuals}{ffms_base}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "ffms_base".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "ffms_base", transforms = c("sigmoid"))
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
