% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.ffms_merged}
\alias{residuals.ffms_merged}
\title{Residuals for FFMS Merged Model}
\usage{
\method{residuals}{ffms_merged}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "ffms_merged".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, 
method = "ffms.parallel", transforms = c("sigmoid"), 
runs = 2, cores = 1)
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
