% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.mjmcmc}
\alias{residuals.mjmcmc}
\title{Residuals for MJMCMC Model}
\usage{
\method{residuals}{mjmcmc}(object, y, x, ...)
}
\arguments{
\item{object}{Object of class "mjmcmc".}

\item{y}{Respnse.}

\item{x}{Covariates.}

\item{...}{Additional arguments (ignored).}
}
\value{
Vector of residuals.
}
\description{
Computes residuals as the difference between observed and predicted values.
}
\examples{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet, method = "mjmcmc")
hist(residuals(model, exoplanet[,1], exoplanet[,-1]))
}
