% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_optimize.R
\name{sgd}
\alias{sgd}
\title{(Batch) (Stochastic) Gradient Descent}
\usage{
sgd(gradient, data = NULL, ctrl = list())
}
\arguments{
\item{gradient}{The gradient of the objective function.}

\item{data}{The dataset to use for the gradient, if applicable, if NULL, no data is used.}

\item{ctrl}{A list of control parameters.}
}
\description{
(Batch) (Stochastic) Gradient Descent
}
