% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_optimize.R
\name{sic.grad}
\alias{sic.grad}
\title{Gradient for GLM with SIC Penalty}
\usage{
sic.grad(x, y, theta, idx, family, lambda, epsilon)
}
\arguments{
\item{x}{A matrix containing the covariates (including an intercept if one wants to use one).}

\item{y}{The dependent variable.}

\item{theta}{The coefficient vector to use.}

\item{idx}{The observation indices to use.}

\item{family}{A glm family for the distribution to use, i.e. "binomial()".}

\item{lambda}{A vector of feature-specific SIC penalties.}

\item{epsilon}{The current epsilon for the epsilon-telescope.}
}
\description{
Gradient for GLM with SIC Penalty
}
