% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_optimize.R
\name{sic.sgd}
\alias{sic.sgd}
\title{(Batch) (Stochastic) Gradient Descent for SIC}
\usage{
sic.sgd(x, y, family, lambda, epsilon, sgd.ctrl = NULL)
}
\arguments{
\item{x}{A matrix containing the covariates.}

\item{y}{The dependent variable.}

\item{family}{A glm family for the distribution to use.}

\item{lambda}{A vector of feature-specific SIC penalties.}

\item{epsilon}{The current epsilon for the epsilon-telescope.}

\item{sgd.ctrl}{control parameters to pass to the sgd function.}
}
\description{
(Batch) (Stochastic) Gradient Descent for SIC
}
