% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_optimize.R
\name{sic_optimize.loop}
\alias{sic_optimize.loop}
\title{SIC optimize loop (replaces mjmcmc.loop)}
\usage{
sic_optimize.loop(
  data.t,
  complex,
  loglik.pi,
  model.cur,
  N.this,
  probs,
  params,
  sub,
  verbose
)
}
\description{
SIC optimize loop (replaces mjmcmc.loop)
}
