% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{summary.ffms_predict}
\alias{summary.ffms_predict}
\title{Summary of FFMS Prediction Object}
\usage{
\method{summary}{ffms_predict}(object, ...)
}
\arguments{
\item{object}{Object of class "ffms_predict".}

\item{...}{Additional arguments (ignored).}
}
\value{
Prints a summary and returns NULL.
}
\description{
Provides a detailed summary of an FFMS prediction object, including prediction ranges.
}
\examples{
\donttest{
data(exoplanet)
model <- ffms(semimajoraxis ~ ., data = exoplanet)
pred <- predict(model, exoplanet[51:60, -1])
summary(pred)
}
}
