% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fms_parallel}
\alias{summary.fms_parallel}
\title{Function to Print a Quick Summary of the Results}
\usage{
\method{summary}{fms_parallel}(object, top_N = 10, labels = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{object}{The results to use}

\item{top_N}{The number of top unique models to report. Defaults to 10.}

\item{labels}{Should the covariates be named, or just referred to as their place in the data.frame.}

\item{verbose}{If the summary should be printed to the console or just returned, defaults to TRUE}

\item{...}{Not used.}
}
\description{
Function to Print a Quick Summary of the Results
}
