% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{ffms.parallel}
\alias{ffms.parallel}
\title{Run Multiple FFMS (Genetically Modified FMS) Runs in Parallel.}
\usage{
ffms.parallel(
  x,
  y,
  loglik.pi = NULL,
  mlpost_params = list(family = "gaussian", beta_prior = list(type = "g-prior")),
  loglik.alpha = gaussian.loglik.alpha,
  transforms,
  pop.max = 15,
  penalty_a = 1,
  prob_filter = 0.6,
  prob_gen = c(0.4, 0.4, 0.1, 0.1),
  runs = 2,
  cores = getOption("mc.cores", 2L),
  verbose = FALSE,
  merge.options = list(populations = "best", complex.measure = 2, tol = 1e-07),
  ...
)
}
\arguments{
\item{x}{matrix containing the design matrix with data to use in the algorithm}

\item{y}{response variable}

\item{loglik.pi}{The (log) density to explore}

\item{mlpost_params}{parameters for the estimator function loglik.pi}

\item{loglik.alpha}{The likelihood function to use for alpha calculation}

\item{transforms}{A Character vector including the names of the non-linear functions to be used by the modification}

\item{pop.max}{Maximum population size. Default is 15.}

\item{penalty_a}{Penalty parameter for features. Default is 1.}

\item{prob_filter}{Probability for filtering. Default is 0.6.}

\item{prob_gen}{Probabilities for generation operators. Default is c(0.4, 0.4, 0.1, 0.1).}

\item{runs}{The number of runs to run}

\item{cores}{The number of cores to run on}

\item{verbose}{A logical denoting if messages should be printed}

\item{merge.options}{A list of options to pass to the \code{\link[=merge_results]{merge_results()}} function run after the run}

\item{...}{Further parameters passed to fms_base.}
}
\value{
Results from multiple ffms_base runs
}
\description{
Run Multiple FFMS (Genetically Modified FMS) Runs in Parallel.
}
\examples{
result <- ffms.parallel(
  runs = 1,
  cores = 1,
  loglik.pi = NULL,
  y = matrix(rnorm(100), 100),
  x = matrix(rnorm(600), 100),
  transforms = c("p0", "exp_dbl")
)

summary(result)

plot(result)

}
